+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Frameworks:
 
 OS:
 - LilyGo T-Watch S3 Plus: enable audio input (PDM microphone) and output (I2S speaker)
+- LilyGo T-Watch S3 Plus: enable battery-powered Real Time Clock to keep time